token&userinfo

+++ src/main/java/com/sparta/crud/service/MemoService.java
@@ -4,11 +4,17 @@ package com.sparta.crud.service;
 import com.sparta.crud.dto.MemoRequestDto;
 import com.sparta.crud.dto.MemoResponseDto;
 import com.sparta.crud.entity.Memo;
+import com.sparta.crud.entity.User;
+import com.sparta.crud.entity.UserRoleEnum;
+import com.sparta.crud.jwt.JwtUtil;
 import com.sparta.crud.repository.MemoRepository;
+import com.sparta.crud.repository.UserRepository;
+import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,6 +23,9 @@ import java.util.List;
 public class MemoService {
 
     private final MemoRepository memoRepository;
+    private final UserRepository userRepository;
+    private final JwtUtil jwtUtil;
+
 
 /*    @Transactional
     public Memo createMemo(MemoRequestDto requestDto) {
@@ -25,11 +34,56 @@ public class MemoService {
         return memo;
     }*/
 
+    //    @Transactional
+//    public Memo createMemo(MemoRequestDto requestDto, HttpServletRequest request) {
+//        // Request에서 Token 가져오기
+//        String token = jwtUtil.resolveToken(request);
+//        Claims claims;
+//
+//        if (token != null) {
+//            // Token 검증
+//            if (jwtUtil.validateToken(token)) {
+//                // 토큰에서 사용자 정보 가져오기
+//                claims = jwtUtil.getUserInfoFromToken(token);
+//            } else {
+//                throw new IllegalArgumentException("Token Error");
+//            }
+//            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+//                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+//            );
+//
+//            Memo memo = new Memo(requestDto);
+//            memoRepository.save(memo);
+//            return memo;
+//        } else {
+//            return null;
+//        }
+//    }
     @Transactional
-    public Memo createMemo(MemoRequestDto requestDto) {
-        Memo memo = new Memo(requestDto);
-        memoRepository.save(memo);
-        return memo;
+    public Memo createMemo(MemoRequestDto requestDto, HttpServletRequest request) {
+        // Import Token from Request
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+
+        if (token != null) {
+            // Token verification
+            if (jwtUtil.validateToken(token)) {
+                // Get user information from token
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("User does not exist")
+            );
+
+            Memo memo = new Memo(requestDto);
+            memo.setUser(user);
+            memoRepository.save(memo);
+            return memo;
+        } else {
+            throw new IllegalArgumentException("Token Error");
+        }
     }
 
 
@@ -40,7 +94,7 @@ public class MemoService {
 
     @Transactional(readOnly = true)
     public List<MemoResponseDto> getMemos() {
-        List<Memo> memos = memoRepository.findAllByOrderByModifiedAtDesc();
+        List<Memo> memos = memoRepository.findAllByOrderByCreatedAtDesc();
         List<MemoResponseDto> memoResponseDtos = new ArrayList<>();
         for (Memo memo : memos) {
             memoResponseDtos.add(new MemoResponseDto(memo));
@@ -55,17 +109,74 @@ public class MemoService {
     }
 
     @Transactional
-    public Long update(Long id, MemoRequestDto requestDto) {
+    public Long update(Long id, MemoRequestDto requestDto, HttpServletRequest request) {
+
+        // Request에서 Token 가져오기
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
         Memo memo = memoRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
-        );  //db에 내가 수정할 메모가 있는지 확인
-        memo.update(requestDto);
-        return memo.getId();
+        );
+
+        if (token != null) {
+            // Token 검증
+            if (jwtUtil.validateToken(token)) {
+                // 토큰에서 사용자 정보 가져오기
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+            );
+            if (memo.getUser().getId() == user.getId() || user.getRole().equals(UserRoleEnum.ADMIN)) {
+                memo.update(requestDto);
+                return memo.getId();
+            }else {
+                return null;
+            }
+//
+//        Memo memo = memoRepository.findById(id).orElseThrow(
+//                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+//        );  //db에 내가 수정할 메모가 있는지 확인
+
+
+
+        } else {
+            return null;
+        }
     }
 
     @Transactional
-    public Long deleteMemo(Long id) {
-        memoRepository.deleteById(id);
-        return id;
+    public Long deleteMemo(Long id, HttpServletRequest request) {
+        // Request에서 Token 가져오기
+        String token = jwtUtil.resolveToken(request);
+        Claims claims;
+
+        Memo memo = memoRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("게시글 없음")
+        );
+
+        if (token != null) {
+            // Token 검증
+            if (jwtUtil.validateToken(token)) {
+                // 토큰에서 사용자 정보 가져오기
+                claims = jwtUtil.getUserInfoFromToken(token);
+            } else {
+                throw new IllegalArgumentException("Token Error");
+            }
+            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
+                    () -> new IllegalArgumentException("사용자가 존재하지 않습니다.")
+            );
+
+            if (memo.getUser().getId() == user.getId() || user.getRole().equals(UserRoleEnum.ADMIN)) {
+                memoRepository.deleteById(id);
+                return id;
+            }else {
+                return null;
+            }
+        } else {
+            return null;
+        }
     }
 }
